WeightedVertex extends Vertex
